+++ src/main/java/com/kainos/ea/backend/controllers/BandCompetencyController.java
@@ -16,6 +16,10 @@ public class BandCompetencyController {
     @Autowired
     BandCompetencyService bandCompetencyService;
 
+    public BandCompetencyController(BandCompetencyService service) {
+        this.bandCompetencyService = service;
+    }
+
     @GetMapping("/competency/{bandName}")
     public @ResponseBody
     Iterable<BandCompetency> getCompetenciesByBand(@PathVariable("bandName") String bandName){
