+++ src/test/java/com/kainos/ea/backend/BackendApplicationTests.java
@@ -2,8 +2,10 @@ package com.kainos.ea.backend;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.TestPropertySource;
 
 @SpringBootTest
+@TestPropertySource(locations = "/application-test.properties")
 class BackendApplicationTests {
 
 	@Test
