--- src/test/java/com/kainos/ea/backend/controllers/CompetencyControllerTest.java
@@ -1,35 +1,0 @@
-package com.kainos.ea.backend.controllers;
-
-import org.junit.jupiter.api.Test;
-import org.skyscreamer.jsonassert.JSONAssert;
-import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.ResponseEntity;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class CompetencyControllerTest {
-
-    @LocalServerPort
-    private final int port = 8080;
-
-    TestRestTemplate restTemplate = new TestRestTemplate();
-    HttpHeaders headers = new HttpHeaders();
-
-    @Test
-    public void getCompetenciesByBandHappyPath() throws Exception {
-        HttpEntity<String> entity = new HttpEntity<String>(null, headers);
-        ResponseEntity<String> response = restTemplate.exchange(createURLWithPort("/competency/test"), HttpMethod.GET, entity, String.class);
-        String expected = "[{\"competencyName\":\"test\"}]";
-
-        JSONAssert.assertEquals(expected, response.getBody(), false);
-    }
-
-    private String createURLWithPort(String uri) {
-
-        return "http://localhost:" + port + uri;
-    }
-}
