+++ src/test/resources/application-test.properties
@@ -1,0 +1,3 @@
+spring.datasource.url = jdbc:h2:mem:test
+spring.datasource.driver-class-name=org.h2.Driver
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.H2Dialect
